- Fix I18N on change pages. - Fix retrieve clients on ClientConverter

+++ service-order/service-order-ejb/src/main/java/br/com/codeshare/data/PhoneRepository.java
@@ -10,7 +10,7 @@ import br.com.codeshare.model.Phone;
 @RequestScoped
 public class PhoneRepository extends AbstractRepository<Phone>{
 	
-	public List<Phone> findClientPhone(Integer clientId) {
+	public List<Phone> findClientPhone(Long clientId) {
 		TypedQuery<Phone> query = 
 				em.createQuery("from Phone phone where phone.client.id = :clientId",Phone.class);
 		query.setParameter("clientId", clientId);
